// src/types/vehicle.ts
export interface Vehicle {
  name: string;
  model: string;
  cargo_capacity: string;
  consumables: string;
  cost_in_credits: string;
  crew: string;
  length: string;
  manufacturer: string;
  max_atmosphering_speed: string;
  vehicle_class: string;
  passengers: string;
  pilots: string[];
  films: string[];
  created: string;
  edited: string;
  url: string;
}